Continue tab complete on single hit results into result children
Tab complete in the lua prompt shows . on tables, but can give the wrong impression that there are no keys. Consider that the user types =filesyste[tab] and then sees =filesystem. and presses tab again. The hint table has only one entry, "=filesystem." and is cycling through that one entry and not reconsidering children of filesystem .. which is what the prompt makes the user think. This change injects a metamethod lookup in the 2nd index of a single result to repop the hint list with new children.