optimizations for install and cp and filesystem links fixes
optmized code and split up install to reduce memory costs. Can now install on 1x T1 RAM systems.

found old bug in filesystem:
touch a
ln a b
mkdir d
cd d
touch b
rm b

And the link in the parent dir would be lost. The reason is that the virtual nodes created for the link were being selected when removing the b file in d, as well as removing b from the real fs. The fix was to check that the node was the correct node, and not just remove the key from any node returned from findNode